Convert resource state tests to molecule
