+++ resource-state/build.gradle.kts
@@ -16,6 +16,7 @@
 
 import com.alexvanyo.composelife.buildlogic.FormFactor
 import com.alexvanyo.composelife.buildlogic.configureGradleManagedDevices
+import com.alexvanyo.composelife.buildlogic.jvmMolecule
 
 plugins {
     alias(libs.plugins.convention.kotlinMultiplatform)
@@ -39,6 +40,7 @@ android {
 kotlin {
     androidTarget()
     jvm("desktop")
+    jvmMolecule(this)
 
     sourceSets {
         val commonMain by getting {
@@ -64,27 +66,24 @@ kotlin {
         val commonTest by getting {
             dependencies {
                 implementation(libs.kotlinx.coroutines.test)
+                implementation(libs.molecule)
                 implementation(libs.turbine)
-                implementation(projects.kmpAndroidRunner)
             }
         }
-        val jbTest by creating {
+        val jvmTest by creating {
             dependsOn(commonTest)
-            dependencies {
-                implementation(libs.jetbrains.compose.uiTestJunit4)
-            }
+        }
+        val moleculeTest by getting {
+            dependsOn(jvmTest)
+        }
+        val jbTest by creating {
+            dependsOn(jvmTest)
         }
         val desktopTest by getting {
             dependsOn(jbTest)
         }
         val androidSharedTest by getting {
             dependsOn(jbTest)
-            dependencies {
-                implementation(libs.androidx.compose.uiTestJunit4)
-                implementation(libs.androidx.test.core)
-                implementation(libs.androidx.test.espresso)
-                implementation(projects.testActivity)
-            }
         }
     }
 }
